Read badge import PersonId as long so personnummer keys do not crash
v1 sometimes keyed Person on a personnummer (e.g. 200806048724) instead of the Scoutnet member number. BadgePartDoneImport and BadgeCompletedImport declared PersonId as int, so System.Text.Json threw while deserializing and aborted the whole import at step 11.

Both already guard with 'PersonId <= 0 || PersonId > int.MaxValue', but an int can never exceed int.MaxValue, so that check was unreachable. Widening to long matches MeetingAttendanceImport and lets the guard skip the rows as intended.

Note TroopPersonImport.PersonId is still int. It survives only because transform_data.py filters those rows out beforehand.

+++ src/Skojjt.Infrastructure/Services/DataMigrationService.cs
@@ -1197,8 +1197,11 @@ public record TroopBadgeImport(
     int? SortOrder
     );
 
+// PersonId is long, as in MeetingAttendanceImport: v1 sometimes keyed Person on
+// a personnummer (e.g. 200806048724), which overflows Int32. The import filters
+// those out, but only if deserialization gets far enough to run the check.
 public record BadgePartDoneImport(
-    int PersonId,
+    long PersonId,
     int BadgeId,
     int PartIndex,
     bool? IsScoutPart,
@@ -1207,7 +1210,7 @@ public record BadgePartDoneImport(
 );
 
 public record BadgeCompletedImport(
-    int PersonId,
+    long PersonId,
     int BadgeId,
     string? Examiner,
     string? CompletedDate
